Add files via upload
Portfolio Website
A modern and responsive portfolio website built using HTML, CSS, and JavaScript. The website showcases my skills, experience, and achievements in a visually appealing and interactive way.

Design and Style
The website uses a clean and minimalist design, with a focus on typography and white space. The color palette is based on shades of blue and gray, with accent colors used sparingly to draw attention to important elements.The website uses the Poppins font family and a color palette of black, white, and shades of blue, with a main color of #00ffee.The primary font used throughout the website is "Roboto", with "Sans-Serif" as a fallback. Headings use a bold and all-caps style to create visual hierarchy, while body text is set in a comfortable reading size and line height.The website also includes testimonials from satisfied clients and a contact form for potential employers or clients to get in touch in an easy way.

Functionality
The website features smooth scrolling and subtle animations and transitions to enhance the user experience. The navigation menu is sticky and collapses into a hamburger icon on smaller screens.

The homepage features a full-screen hero section with a background video and a welcoming message. The portfolio section showcases my projects in a grid layout with hover effects, and the about section provides an overview of my background, skills, and achievements.

About Me
Proficient in HTML, CSS, JavaScript, and ReactJS, I build responsive and interactive web applications. Experienced in Python and C++ for backend development and machine learning, leveraging TensorFlow and PyTorch for deep learning models. Strong background in data preprocessing, feature engineering, and real-time system integration.